Implement JavaFX components
Created a SnakeGame class, an instance of which will represent the game
in progress. An instance of MosaicCanvas is used as the board, and the
instance of Snake as the snake.

Added keyPressed handlers for testing.
Added method to draw snake completely.
Added method to update visual representation of snake on Canvas
which only updates the new head and old tail; more effcient than
redrawing entire canvas, and will be needed once obstacles are
added in.

Issues:
1. Inital motion with keypressed leaves behind one green square
2. Need to implement logic ensuring snake does not go off board
   within Snake class
3. Need to implement animation with AnimationTimer
4. Would like a way to adjust game speed without changing framerate.