Resolve issue with snake not turning at wall
Problem was that at bottom and right hand walls, the snake was not
turning and moved straight through. Turns out the boardWidth and
boardHeight fields were double values representing canvas size
in pixels, rather than number of columns and rows. That has been changed
and the turning works correctly when the snake runs into a wall.

Replace KeyHandle() with KeyTurn()

New method, that updates the direction of the snake, with some
restrictions:
- only allows 90 degree turns,
- and does not allow turning into a wall.
No longer changes the snake position itself, only sets the direction,
and then subsequently updatePosForNewFrame() will update the position.
This removes some redundant code.